+++ application/config/routes.php
@@ -117,6 +117,7 @@ if(is_cli()) {
 	$route['admin/update_series']        = 'AdminCLI/updateSeries';
 	$route['admin/update_series_custom'] = 'AdminCLI/updateSeriesCustom';
 	$route['admin/update_titles']        = 'AdminCLI/updateTitles';
+	$route['admin/refollow_custom']      = 'AdminCLI/refollowCustom';
 	$route['admin/test']                 = 'AdminCLI/test';
 }